+++ HabboHotel/Items/Wired/IWiredExecutable.cs
@@ -4,3 +4,18 @@ public interface IWiredExecutable
 {
     bool Execute(WiredExecutionContext context);
 }
+
+internal interface IWiredChatExecutable
+{
+    bool Execute(WiredChatExecutionContext context);
+}
+
+internal interface IWiredActorItemExecutable
+{
+    bool Execute(WiredActorItemExecutionContext context);
+}
+
+internal interface IWiredEmptyExecutable
+{
+    bool Execute(WiredEmptyExecutionContext context);
+}